fix: centralize order runtime helpers and failure logging
