+++ internal/api/handler_orders.go
@@ -3,11 +3,8 @@ package api
 import (
 	"errors"
 	"fmt"
-	"sort"
 	"strings"
-	"time"
 
-	"github.com/gastownhall/gascity/internal/beads"
 	"github.com/gastownhall/gascity/internal/orders"
 )
 
@@ -93,164 +90,6 @@ func toOrderResponse(a orders.Order) orderResponse {
 		Rig:           a.Rig,
 		CaptureOutput: a.IsExec(), // exec orders capture output
 	}
-}
-
-func beadLastRunFunc(store beads.Store) orders.LastRunFunc {
-	return func(name string) (time.Time, error) {
-		if store == nil {
-			return time.Time{}, nil
-		}
-		label := "order-run:" + name
-		results, err := store.List(beads.ListQuery{
-			Label:         label,
-			Limit:         1,
-			IncludeClosed: true,
-			Sort:          beads.SortCreatedDesc,
-		})
-		if err != nil {
-			return time.Time{}, err
-		}
-		if len(results) == 0 {
-			return time.Time{}, nil
-		}
-		return results[0].CreatedAt, nil
-	}
-}
-
-// orderStoresForState returns the primary store for an order plus the
-// city store fallback when available. Rig-scoped orders read rig first so
-// the current store wins over any legacy city fallback.
-func orderStoresForState(state State, a orders.Order) ([]beads.Store, error) {
-	stores := make([]beads.Store, 0, 2)
-	if strings.TrimSpace(a.Rig) == "" {
-		store := state.CityBeadStore()
-		if store == nil {
-			return nil, fmt.Errorf("city bead store is unavailable")
-		}
-		return []beads.Store{store}, nil
-	}
-
-	rigStore := state.BeadStore(a.Rig)
-	if rigStore == nil {
-		return nil, fmt.Errorf("rig %q bead store is unavailable", a.Rig)
-	}
-	stores = append(stores, rigStore)
-
-	if cityStore := state.CityBeadStore(); cityStore != nil {
-		stores = append(stores, cityStore)
-	}
-	return stores, nil
-}
-
-// orderLastRunFnAcrossStores returns the most recent run time across a set
-// of stores for a single order name.
-func orderLastRunFnAcrossStores(stores ...beads.Store) orders.LastRunFunc {
-	return func(name string) (time.Time, error) {
-		var latest time.Time
-		for _, store := range stores {
-			if store == nil {
-				continue
-			}
-			last, err := beadLastRunFunc(store)(name)
-			if err != nil {
-				return time.Time{}, err
-			}
-			if last.After(latest) {
-				latest = last
-			}
-		}
-		return latest, nil
-	}
-}
-
-// orderCursorFuncAcrossStores merges seq cursors from multiple stores.
-func orderCursorFuncAcrossStores(stores ...beads.Store) orders.CursorFunc {
-	return func(name string) uint64 {
-		label := "order-run:" + name
-		var latest uint64
-		for _, store := range stores {
-			if store == nil {
-				continue
-			}
-			results, err := store.List(beads.ListQuery{
-				Label:         label,
-				Limit:         10,
-				IncludeClosed: true,
-				Sort:          beads.SortCreatedDesc,
-			})
-			if err != nil || len(results) == 0 {
-				continue
-			}
-			labelSets := make([][]string, 0, len(results))
-			for _, b := range results {
-				labelSets = append(labelSets, b.Labels)
-			}
-			if seq := orders.MaxSeqFromLabels(labelSets); seq > latest {
-				latest = seq
-			}
-		}
-		return latest
-	}
-}
-
-// orderHistoryBeadsAcrossStores merges order history rows from a primary
-// store and its fallback stores while preserving recency ordering.
-func orderHistoryBeadsAcrossStores(stores []beads.Store, scopedName string) ([]beads.Bead, error) {
-	label := "order-run:" + scopedName
-	seen := make(map[string]bool)
-	results := make([]beads.Bead, 0)
-
-	for i, store := range stores {
-		if store == nil {
-			continue
-		}
-		rows, err := store.List(beads.ListQuery{
-			Label:         label,
-			IncludeClosed: true,
-			Sort:          beads.SortCreatedDesc,
-		})
-		if err != nil {
-			if i == 0 {
-				return nil, err
-			}
-			continue
-		}
-		for _, row := range rows {
-			if seen[row.ID] {
-				continue
-			}
-			seen[row.ID] = true
-			results = append(results, row)
-		}
-	}
-
-	sort.SliceStable(results, func(i, j int) bool {
-		return results[i].CreatedAt.After(results[j].CreatedAt)
-	})
-	return results, nil
-}
-
-// orderHistoryBeadAcrossStores finds a bead by ID across a primary store
-// and its fallbacks.
-func orderHistoryBeadAcrossStores(stores []beads.Store, beadID string) (beads.Bead, error) {
-	var lastErr error
-	for _, store := range stores {
-		if store == nil {
-			continue
-		}
-		bead, err := store.Get(beadID)
-		if err == nil {
-			return bead, nil
-		}
-		if errors.Is(err, beads.ErrNotFound) {
-			continue
-		}
-		lastErr = err
-	}
-	if lastErr != nil {
-		return beads.Bead{}, lastErr
-	}
-	return beads.Bead{}, beads.ErrNotFound
 }
 
 // lastRunOutcomeFromLabels extracts the run outcome from bead labels.
